+++ internal/bundler/bundler_dce_test.go
@@ -1829,6 +1829,74 @@ func TestDCETypeOfEqualsStringGuardCondition(t *testing.T) {
 	})
 }
 
+func TestDCETypeOfCompareStringGuardCondition(t *testing.T) {
+	dce_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/entry.js": `
+				// Everything here should be removed as dead code due to tree shaking
+				var REMOVE_1 = typeof x <= 'u' ? x : null
+				var REMOVE_1 = typeof x < 'u' ? x : null
+				var REMOVE_1 = typeof x >= 'u' ? null : x
+				var REMOVE_1 = typeof x > 'u' ? null : x
+				var REMOVE_1 = typeof x <= 'u' && x
+				var REMOVE_1 = typeof x < 'u' && x
+				var REMOVE_1 = typeof x >= 'u' || x
+				var REMOVE_1 = typeof x > 'u' || x
+				var REMOVE_1 = 'u' >= typeof x ? x : null
+				var REMOVE_1 = 'u' > typeof x ? x : null
+				var REMOVE_1 = 'u' <= typeof x ? null : x
+				var REMOVE_1 = 'u' < typeof x ? null : x
+				var REMOVE_1 = 'u' >= typeof x && x
+				var REMOVE_1 = 'u' > typeof x && x
+				var REMOVE_1 = 'u' <= typeof x || x
+				var REMOVE_1 = 'u' < typeof x || x
+
+				// Everything here should be kept as live code because it has side effects
+				var keep_1 = typeof x <= 'u' ? y : null
+				var keep_1 = typeof x < 'u' ? y : null
+				var keep_1 = typeof x >= 'u' ? null : y
+				var keep_1 = typeof x > 'u' ? null : y
+				var keep_1 = typeof x <= 'u' && y
+				var keep_1 = typeof x < 'u' && y
+				var keep_1 = typeof x >= 'u' || y
+				var keep_1 = typeof x > 'u' || y
+				var keep_1 = 'u' >= typeof x ? y : null
+				var keep_1 = 'u' > typeof x ? y : null
+				var keep_1 = 'u' <= typeof x ? null : y
+				var keep_1 = 'u' < typeof x ? null : y
+				var keep_1 = 'u' >= typeof x && y
+				var keep_1 = 'u' > typeof x && y
+				var keep_1 = 'u' <= typeof x || y
+				var keep_1 = 'u' < typeof x || y
+
+				// Everything here should be kept as live code because it has side effects
+				var keep_2 = typeof x <= 'u' ? null : x
+				var keep_2 = typeof x < 'u' ? null : x
+				var keep_2 = typeof x >= 'u' ? x : null
+				var keep_2 = typeof x > 'u' ? x : null
+				var keep_2 = typeof x <= 'u' || x
+				var keep_2 = typeof x < 'u' || x
+				var keep_2 = typeof x >= 'u' && x
+				var keep_2 = typeof x > 'u' && x
+				var keep_2 = 'u' >= typeof x ? null : x
+				var keep_2 = 'u' > typeof x ? null : x
+				var keep_2 = 'u' <= typeof x ? x : null
+				var keep_2 = 'u' < typeof x ? x : null
+				var keep_2 = 'u' >= typeof x || x
+				var keep_2 = 'u' > typeof x || x
+				var keep_2 = 'u' <= typeof x && x
+				var keep_2 = 'u' < typeof x && x
+			`,
+		},
+		entryPaths: []string{"/entry.js"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			OutputFormat:  config.FormatIIFE,
+			AbsOutputFile: "/out.js",
+		},
+	})
+}
+
 // These unused imports should be removed since they aren't used, and removing
 // them makes the code shorter.
 func TestRemoveUnusedImports(t *testing.T) {